Exacted config drawer component
