+++ frontend/src/components/notes-page/NotesPage.tsx
@@ -3,17 +3,11 @@ import './NotesPage.scss';
 import { useTranslator } from '../../service/TranslateService';
 import { DetailNote, DetailNoteQuestion, LoadStatus } from '../../models';
 import {
-	AppDispatch,
 	getDetailNotes,
-	hideNotesColumn,
-	showNotesColumn,
 	useAppDispatch,
 	useAppSelector
 } from '../../state';
 import {
-	Checkbox,
-	Drawer,
-	FormControlLabel,
 	Icon,
 	IconButton,
 	Table,
@@ -24,6 +18,7 @@ import {
 	TableRow,
 	Typography
 } from '@mui/material';
+import NotesTableConfigDrawer from './notes-table-config-drawer/NotesTableConfigDrawer';
 
 export default function NotesPage() {
 	const translate = useTranslator();
@@ -121,70 +116,5 @@ export default function NotesPage() {
 				handleClose={ () => setConfigDrawerOpen(false) }
 			/>
 		</div>
-	);
-}
-
-function NotesTableConfigDrawer(props: {
-	isOpen: boolean;
-	handleClose: () => void;
-}) {
-	const dispatch: AppDispatch = useAppDispatch();
-	const translate = useTranslator();
-	const columns: string[] = useAppSelector(state => state.detailNotes.questionNames);
-	const hiddenColumns: string[] = useAppSelector(state => state.detailNotes.hiddenQuestionNames);
-
-	const _hideColumn = (column: string): void => {
-		dispatch(hideNotesColumn(column));
-	};
-
-	const _showColumn = (column: string): void => {
-		dispatch(showNotesColumn(column));
-	};
-
-	const handleCheckboxClick = (column: string, isChecked: boolean): void => {
-		if (isChecked) {
-			_showColumn(column);
-			return;
-		}
-
-		_hideColumn(column);
-	};
-
-	const checkboxes = columns.map((title: string) => (
-		<FormControlLabel
-			key={ title }
-			label={ translate(title) }
-			control={
-				<Checkbox
-					checked={ !hiddenColumns.includes(title) }
-					onChange={ (event) => handleCheckboxClick(title, event.target.checked) }
-				/>
-			}
-		/>
-	));
-
-	return (
-		<Drawer
-			anchor="right"
-			open={ props.isOpen }
-			onClose={ props.handleClose }
-		>
-			<div className="notes-page-config-drawer__header">
-				<Typography variant="h6">Show columns</Typography>
-			</div>
-			<div className="notes-page-config-drawer__body">
-				<FormControlLabel
-					key="TEAM_NUMBER"
-					label={ translate('TEAM_NUMBER') }
-					control={
-						<Checkbox
-							checked={true}
-							disabled={true}
-						/>
-					}
-				/>
-				{ checkboxes }
-			</div>
-		</Drawer>
 	);
 }
